src/Firmware/requirements.txt
pycryptodome==3.10.1
colorama==0.4.3
pybluez==0.23